Modificacion de envio, dataset y relaciones del mismo, imlementacion del login

+++ src/main/java/co/edu/uniquindio/unimarket/servicios/interfaces/CalificacionServicio.java
@@ -7,7 +7,7 @@ public interface CalificacionServicio {
     int crearCalificacion(CalificacionDTO calificacionDTO) throws Exception;
 
 
-    int promedioCalificacion(int idProducto);
+    float promedioCalificacion(int idProducto);
 
 
 }
